[ALPS07354051] GPU: PMU record on r38p1 with LTR and MET
1. Fix pmu record with LTR and MET
2. Fix build ko fail on r38p0

MTK-Commit-Id: 947c30bda53e195e54e925c00ca9f466e84ea0cc

CR-Id: ALPS07354051
Feature: SWPM Kernel
(cherry picked from commit ca8865ad86189f9d21f7fe6c483b985da57e3d31)

+++ gpu_mali/mali_valhall/mali-r38p1/drivers/gpu/arm/midgard/platform/mtk_platform_common/mtk_gpu_trace.h
@@ -14,19 +14,19 @@
 
 
 TRACE_EVENT(perf_index_gpu,
-	TP_PROTO(u32 *gpu_data, u32 lens),
+	TP_PROTO(u64 *gpu_data, u32 lens),
 	TP_ARGS(gpu_data, lens),
 	TP_STRUCT__entry(
-		__dynamic_array(u32, gpu_data, lens)
+		__dynamic_array(u64, gpu_data, lens)
 		__field(u32, lens)
 	),
 	TP_fast_assign(
 		memcpy(__get_dynamic_array(gpu_data), gpu_data,
-			lens * sizeof(u32));
+			lens * sizeof(u64));
 		__entry->lens = lens;
 	),
 	TP_printk("data=%s", __print_array(__get_dynamic_array(gpu_data),
-		__entry->lens, sizeof(u32)))
+		__entry->lens, sizeof(u64)))
 );
 
 #endif /*_MTK_GPU_TRACE_H */
